+++ contracts/interfaces/curve/IFundsModule.sol
@@ -78,6 +78,14 @@ interface IFundsModule {
     function calculatePoolEnter(uint256 lAmount) external view returns(uint256);
 
     /**
+     * @notice Calculates how many pTokens should be given to user for increasing liquidity
+     * @param lAmount Amount of liquid tokens which will be put into the pool
+     * @param liquidityCorrection Amount of liquid tokens to remove from liquidity because it was "virtually" withdrawn
+     * @return Amount of pToken which should be sent to sender
+     */
+    function calculatePoolEnter(uint256 lAmount, uint256 liquidityCorrection) external view returns(uint256);
+    
+    /**
      * @notice Calculates how many pTokens should be taken from user for decreasing liquidity
      * @param lAmount Amount of liquid tokens which will be removed from the pool
      * @return Amount of pToken which should be taken from sender
@@ -97,6 +105,14 @@ interface IFundsModule {
      * @return Amount of pToken which should be taken from sender
      */
     function calculatePoolExitWithFee(uint256 lAmount) external view returns(uint256);
+
+    /**
+     * @notice Calculates amount of pTokens which should be burned/locked when liquidity removed from pool
+     * @param lAmount Amount of liquid tokens beeing withdrawn. Does NOT include part for pool
+     * @param liquidityCorrection Amount of liquid tokens to remove from liquidity because it was "virtually" withdrawn
+     * @return Amount of pTokens to burn/lock
+     */
+    function calculatePoolExitWithFee(uint256 lAmount, uint256 liquidityCorrection) external view returns(uint256);
 
     /**
      * @notice Current pool liquidity