+++ contracts/test/FundsModuleStub.sol
@@ -69,11 +69,19 @@ contract FundsModuleStub is Module, IFundsModule {
         return 0;
     }
 
+    function calculatePoolEnter(uint256, uint256) external view returns(uint256) {
+        return 0;
+    }
+
     function calculatePoolExit(uint256) external view returns(uint256) {
         return 0;
     }
     
     function calculatePoolExitWithFee(uint256) external view returns(uint256){
+        return 0;
+    }
+    
+    function calculatePoolExitWithFee(uint256, uint256) external view returns(uint256){
         return 0;
     }